# public/swagger.yaml
openapi: 3.0.3
info:
  title: Vehicle Incident Management API
  description: API documentation for managing vehicle incidents
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local server
paths:
  /incidents:
    get:
      summary: Get all incidents
      responses:
        "200":
          description: A list of incidents with pagination metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentsResponse"
    post:
      summary: Create a new incident
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/IncidentCreateInput"
      responses:
        "201":
          description: The created incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Validation failed: invalid status"
...
  /incidents/{id}/updates:
    post:
      summary: Add a comment/update to an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - comment
              properties:
                userId:
                  type: integer
                  example: 3
                comment:
                  type: string
                  example: "Changed status to IN_PROGRESS"
      responses:
        "201":
          description: The added incident update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentUpdate"
        "400":
          description: Invalid input data
        "404":
          description: Incident not found
...
components:
  schemas:
    Car:
      type: object
      properties:
        id:
          type: integer
          example: 1
        make:
          type: string
          example: "Toyota"
        model:
          type: string
          example: "Corolla"
        licensePlate:
          type: string
          example: "AB123CD"
    User:
      type: object
      properties:
        id:
          type: integer
          example: 7
        name:
          type: string
          example: "John Doe"
    CarReading:
      type: object
      properties:
        id:
          type: integer
          example: 44
        odometerReading:
          type: number
          example: 45230
        reportedAt:
          type: string
          format: date-time
          example: "2025-08-29T09:30:00Z"
...
    IncidentUpdate:
      type: object
      properties:
        userId:
          type: integer
          example: 3
        comment:
          type: string
          example: "Changed status to IN_PROGRESS"
    Incident:
      type: object
      properties:
        id:
          type: integer
          example: 15
        title:
          type: string
          example: "Brake failure"
        description:
          type: string
          example: "Brake not working properly"
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
          default: LOW
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, RESOLVED, CLOSED]
          default: PENDING
        type:
          type: string
          enum:
            - ACCIDENT
            - BREAKDOWN
            - MAINTENANCE_ISSUE
            - TRAFFIC_VIOLATION
            - FUEL_ISSUE
            - THEFT
            - VANDALISM
            - OTHER
        location:
          type: string
          nullable: true
          example: "Highway 21"
        occurredAt:
          type: string
          format: date-time
          example: "2025-08-29T10:00:00Z"
        reportedAt:
          type: string
          format: date-time
          example: "2025-08-29T12:00:00Z"
        estimatedCost:
          type: number
          nullable: true
          example: 5000
        car:
          $ref: "#/components/schemas/Car"
        reportedBy:
          $ref: "#/components/schemas/User"
        assignedTo:
          allOf:
            - $ref: "#/components/schemas/User"
          nullable: true
        carReading:
          allOf:
            - $ref: "#/components/schemas/CarReading"
          nullable: true
    Meta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 200
        totalPages:
          type: integer
          example: 20
    IncidentsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Incident"
        meta:
          $ref: "#/components/schemas/Meta"
    IncidentCreateInput:
      type: object
      required:
        - carId
        - reportedById
        - title
        - description
        - type
        - occurredAt
      properties:
        carId:
          type: integer
          example: 1
        reportedById:
          type: integer
          example: 5
        assignedToId:
          type: integer
          nullable: true
          example: 10
        title:
          type: string
          example: "Brake failure"
        description:
          type: string
          example: "Brake not working properly"
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
          default: LOW
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, RESOLVED, CLOSED]
          default: PENDING
        type:
          type: string
          enum: [ACCIDENT, BREAKDOWN, MAINTENANCE_ISSUE]
        location:
          type: string
          nullable: true
          example: "Highway 21"
        latitude:
          type: number
          nullable: true
          example: 12.9716
        longitude:
          type: number
          nullable: true
          example: 77.5946
        occurredAt:
          type: string
          format: date-time
          example: "2025-08-29T10:00:00Z"
        reportedAt:
          type: string
          format: date-time
          example: "2025-08-29T12:00:00Z"
        estimatedCost:
          type: number
          nullable: true
          example: 5000
        carReadingId:
          type: integer
          nullable: true
          example: 44
        images:
          type: array
          items:
            type: string
...